rpc-lib: coverity fixes
Scan URL:
https://download.gluster.org/pub/gluster/glusterfs/static-analysis/master/glusterfs-coverity/2017-11-10-0f524f07/html/

ID: 9                            (BAD_SHIFT)
ID: 58                           (CHECKED_RETURN)
ID: 98                           (DEAD_CODE)
ID: 249, 250, 251, 252           (MIXED_ENUMS)
ID: 289, 297                     (NULL_RETURNS)
ID: 609, 613, 622, 644, 653, 655 (UNUSED_VALUE)
ID: 432                          (RESOURCE_LEAK)

BUG: 789278

+++ rpc/rpc-lib/src/rpcsvc.h
@@ -514,8 +514,8 @@ rpcsvc_auth_check (rpcsvc_t *svc, char *volname, char *ipaddr);
 extern int
 rpcsvc_transport_privport_check (rpcsvc_t *svc, char *volname, uint16_t port);
 
-#define rpcsvc_request_seterr(req, err)                 (req)->rpc_err = err
-#define rpcsvc_request_set_autherr(req, err)            (req)->auth_err = err
+#define rpcsvc_request_seterr(req, err)         ((req)->rpc_err = (int)(err))
+#define rpcsvc_request_set_autherr(req, err)    ((req)->auth_err = (int)(err))
 
 extern int rpcsvc_submit_vectors (rpcsvc_request_t *req);
